Init
FSD 폴더 아키텍처를 적용했습니다. 일부 어려운 부분은 간략화했고, docs에
FSD 안내 문서를 작성했습니다.
