+++ .eslintrc
@@ -1,28 +1,29 @@
 {
-  "env": {
-    "browser": true,
-    "es6": true,
-    "node": true,
-  },
-  "root": true,
-  "extends": [
-    "airbnb", 
-    "eslint:recommended",
-    "react-app", 
-    "plugin:react/recommended",
-    "plugin:import/recommended",
-    "plugin:jsx-a11y/recommended",
-    "plugin:react-hooks/recommended",
-    "prettier", 
-    "plugin:prettier/recommended"
-  ],
-  "plugins": [
-    "prettier"
-  ],
-  "rules": {
-    "prettier/prettier": ["error", { "endOfLine": "auto" }],
-    "no-restricted-globals": "off",
-    "no-lone-blocks": "off",
-    "no-unused-vars": "off",
-  },
+	"env": {
+		"browser": true,
+		"es6": true,
+		"node": true,
+	},
+	"root": true,
+	"extends": [
+		"airbnb",
+		"eslint:recommended",
+		"react-app",
+		"plugin:react/recommended",
+		"plugin:import/recommended",
+		"plugin:jsx-a11y/recommended",
+		"plugin:react-hooks/recommended",
+		"prettier",
+		"plugin:prettier/recommended",
+	],
+	"plugins": ["prettier"],
+	"rules": {
+		"prettier/prettier": ["error", { "endOfLine": "auto" }],
+		"no-restricted-globals": "off",
+		"no-lone-blocks": "off",
+		"no-unused-vars": "off",
+		"react/react-in-jsx-scope": "off",
+		"react/jsx-uses-react": "off",
+		"import/extensions": "off",
+	},
 }
