+++ docs/FSD
@@ -1,0 +1,9 @@
+app: 전역 컴포넌트가 들어갑니다. router, app, provider 등 뭔가를 감싸는 컴포넌트는 대개 여기에 위치합니다.
+pages: router가 표시하는 각각의 페이지들입니다.
+shared: 각 페이지나 컴포넌트에서 공유해 사용하는 기능들입니다. utility functions, hooks 등이 있습니다.
+components: 각 페이지 등에서 사용하는 기능별 컴포넌트입니다. entities 혹은 features에 속하지만, 이번 프로젝트에선 그대로 사용합니다.
+layouts: 공통내용을 제외하고 변하는 페이지를 뜻합니다. 예를 들어 login과 signup이 같은 레이아웃을 공유한다면, 그 레이아웃은 여기에 속합니다. widgets에 속하지만, 이번 프로젝트에선 그대로 사용합니다.
+assets: 그림 파일 등 외부에서 끌어다 사용해야되는 파일들을 담습니다.
+
+각 폴더에서 내용이 너무 많아질 수 있기 때문에 각 폴더 아래에 하위 폴더를 만들어서 사용하면 됩니다.
+예를 들어 pages 아래에 company, investment 등등의 하위폴더를 만들어서 사용하면 됩니다.
