--- src/App.jsx
@@ -1,15 +1,0 @@
-import GNB from "./components/GNB";
-import { Outlet } from "react-router-dom";
-
-function App() {
-  return (
-    <>
-      <GNB />
-      <div>
-        <Outlet />
-      </div>
-    </>
-  );
-}
-
-export default App;
