+++ src/index.js
@@ -1,7 +1,7 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import Main from "./Main.jsx";
-import "./root.css";
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import Main from './app/Main.jsx';
+import './root.css';
 
-const root = ReactDOM.createRoot(document.getElementById("root"));
+const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(<Main />);
